+++ odoo/local-src/specific_helpdesk/__manifest__.py
@@ -11,6 +11,7 @@
      'helpdesk',
      'mrp',
      'product',
+     'specific_purchase',
      'stock',
  ],
  'website': 'http://www.camptocamp.com',