+++ odoo/local-src/specific_purchase/models/res_partner.py
@@ -9,3 +9,13 @@ class ResPartner(models.Model):
 
     supplier_discount = fields.Float('Supplier discount %')
     is_back_order_accepted = fields.Boolean('Back order accepted')
+
+    supplier_responsible_id = fields.Many2one(
+        comodel_name='res.users',
+        string='Supplier responsible',
+    )
+
+    supplier_substitute_responsible_id = fields.Many2one(
+        comodel_name='res.users',
+        string='Supplier substitute responsible',
+    )